Prevent invalid data is silently accepted
When user wants to set ``Date`` or ``Datetime`` value and missed one of
the fields (day, year, hour or minute) it silently sets it to None
without error messages which input is not valid.

Unfortunately we cannot check that there was missed one of the inputs by
validator because validator has value (``datetime``) which is already
converted.

I suggest that value will be set into None when user lefts empty fields
*day* and *year*.