Prevent invalid data to be silently accepted
When a user wants sets a ``Date`` or ``Datetime`` value and misses
one of the fields (day, year, hour or minute) it is silently set to
None without any 'input is not valid' error message being displayed.

Unfortunately it can not be checked on a validator because the input
value has already been converted to a ``datetime``.

This commit works around it making sure an error is raised and only
convert it to None if both *day* and *year* are left empty.